==> R/repasse-educacao-analise/01_tratamento_dados.Rmd <==
---
title: "01 Tratamento dos Dados"
author: "pedro-hos@outlook.com"
date: "5/3/2022"
---

### Carregando Dados do CSV

```{r}
library(dplyr)
ideb_repasse <- read.csv(file = "../../data/ideb_repasse/fundeb/ideb_repasse-fundeb.csv")
```

### Removendo NAs

```{r}
#ideb_repasse[is.na(ideb_repasse) == TRUE]
ideb_repasse <- ideb_repasse %>% dplyr::filter(ideb_repasse$ANO_IDH > 0 & ideb_repasse$REPASSE_MATRICULA > 0) %>% na.omit()
```

```{r}
ideb_repasse <- ideb_repasse %>% dplyr::mutate(FAIXA_POP = case_when(
                                    (ideb_repasse$POPULACAO <= 5000) ~ "A",
                                    (ideb_repasse$POPULACAO > 5000 &  ideb_repasse$POPULACAO <= 10000) ~ "B",
                                    (ideb_repasse$POPULACAO > 10000 & ideb_repasse$POPULACAO <= 20000) ~ "C",
                                    (ideb_repasse$POPULACAO > 20000 & ideb_repasse$POPULACAO <= 50000) ~ "D",
                                    (ideb_repasse$POPULACAO > 50000 & ideb_repasse$POPULACAO <= 100000) ~ "E",
                                    (ideb_repasse$POPULACAO > 100000 & ideb_repasse$POPULACAO <= 200000) ~ "F",
                                    (ideb_repasse$POPULACAO > 200000) ~ "G"
                                  ))
```

### Convertendo para factor

```{r}
ideb_repasse$ANO_IDEB = as.factor(ideb_repasse$ANO_IDEB)
ideb_repasse$ANO_IDH = as.factor(ideb_repasse$ANO_IDH)
ideb_repasse$ANO_REPASSE = as.factor(ideb_repasse$ANO_REPASSE)
ideb_repasse$ANO_MATRICULA = as.factor(ideb_repasse$ANO_MATRICULA)

ideb_repasse$PERIODO = as.factor(ideb_repasse$PERIODO)
ideb_repasse$UF = as.factor(ideb_repasse$UF)
ideb_repasse$MUNICIPIO = as.factor(ideb_repasse$MUNICIPIO)
ideb_repasse$REDE = as.factor(ideb_repasse$REDE)
ideb_repasse$CODIGO_MUNICIPIO = as.factor(ideb_repasse$CODIGO_MUNICIPIO)
ideb_repasse$FAIXA_POP = as.factor(ideb_repasse$FAIXA_POP)
```

### Separando os dados por Periodo

```{r}
anos_iniciais <- ideb_repasse %>% dplyr::filter(PERIODO =="FUNDAMENTAL_ANOS_INICIAIS" & IDH_EDUCACAO > 0)
anos_finais <- ideb_repasse %>% dplyr::filter(PERIODO=="FUNDAMENTAL_ANOS_FINAIS" & IDH_EDUCACAO > 0)
ensino_medio <- ideb_repasse %>% dplyr::filter(PERIODO=="ENSINO_MEDIO" & IDH_EDUCACAO > 0)

ideb_media <- ideb_repasse %>% 
  group_by(ANO_IDEB,
           ANO_IDH,
           ANO_MATRICULA,
           ANO_REPASSE,
           CODIGO_MUNICIPIO,
           IDHM,
           IDH_EDUCACAO,
           IDH_LONGEVIDADE,
           IDH_RENDA, 
           MUNICIPIO,
           POPULACAO, 
           REGIAO, 
           UF, 
           VALOR_REPASSADO_EDUCACAO, 
           FAIXA_POP) %>% 
  summarise(MATRICULA_SUM = sum(MATRICULA), 
                    REPASSE_MATRICULA = VALOR_REPASSADO_EDUCACAO/MATRICULA_SUM, 
                    IDEB_MEDIA = mean(IDEB_NOTA)) %>%
  distinct()

```

### Salvando em u RData

```{r}
save(ideb_repasse, file = "../../data/ideb_repasse/fundeb/todos_anos.RData")
save(anos_iniciais, file = "../../data/ideb_repasse/fundeb/anos_iniciais.RData")
save(anos_finais,  file = "../../data/ideb_repasse/fundeb/anos_finais.RData")
save(ensino_medio, file = "../../data/ideb_repasse/fundeb/ensino_medio.RData")
save(ideb_media, file = "../../data/ideb_repasse/fundeb/ideb_media.RData")
```
