==> R/repasse-educacao-analise/02_analise_exploratoria.Rmd <==
---
title: "02 Análise Exploratória"
author: "pedro-hos@outlook.com"
date: "5/3/2022"
output:
  pdf_document: default
  html_document:
    df_print: paged
---

```{r, echo=FALSE}
library("PerformanceAnalytics")
library("corrplot")
library("ggpubr")
library("ggplot2")
```


Os seguinte arquivos foram carregados:

* todos_anos.RData: contendo os dados para todos os períodos;
* anos_iniciais.RData: contendo os anos iniciais do ensino fundamental, ou seja de 1º ao 6º ano;
* anos_finais.RData: contendo os anos finais do ensino fundamental, ou seja de 7º ao 9º ano;
* ensino_medio.RData: contendo os dados para o ensino médio, do 1º do ensino médio ao 3º do ensino médio.
*ideb_media.RData: Média das notas do Ideb para os três periodos, respeitando os anos da aplicação e os municípios

```{r, echo=FALSE}
load(file = "../../data/ideb_repasse/todos_anos.RData")
load(file = "../../data/ideb_repasse/anos_iniciais.RData")
load(file = "../../data/ideb_repasse/anos_finais.RData")
load(file = "../../data/ideb_repasse/ensino_medio.RData")
#load(file = "../../data/ideb_repasse/ideb_media.RData")
```

As colunas para todos os períodos se mantém a mesma, como podemos ver a seguir:

```{r}
colnames(ideb_repasse)
```

E também possuem a seguinte estrutura, com variáveis nominais, como o caso da coluna **PERIODO** e valores contínuos como o caso de **IDEB_NOTA**:

```{r}
dplyr::glimpse(ideb_repasse)
```

Os anos das notas do IDEB para os anos iniciais e finais variam de 2011 à 2019, e os valores do repasse foram coletados sempre respeitando **ANO_IDEB + 1**, ou seja, os repasses de 2012 são para o ano da aplicação do Ideb de 2011, e assim por diante. Os indicadores de IDH são os últimos coletados, em 2010.

* Anos Iniciais

```{r}
summary(anos_iniciais)
```

* Anos Finais

```{r}
summary(anos_finais)
```
Já para o Ensino Médio, os anos são de 2017 e 2019.

```{r}
summary(ensino_medio)
```
```{r}
#summary(ideb_media)
```
Recuperando apenas as variáveis numericas:

```{r}
data <- ideb_repasse[, c(6,7,8,9,10,11,14,17,19)]
```

```{r}
sd(data$VALOR_REPASSADO_EDUCACAO)
sd(data$REPASSE_MATRICULA)
sd(data$IDEB_NOTA)
```

```{r}
# Correlação
cor.test(data$VALOR_REPASSADO_EDUCACAO, data$IDEB_NOTA, method = "pearson")
cor.test(data$IDH_EDUCACAO, data$IDEB_NOTA, method = "pearson")
cor.test(data$VALOR_REPASSADO_EDUCACAO, data$IDH_EDUCACAO, method = "pearson")
cor.test(data$VALOR_REPASSADO_EDUCACAO, data$REPASSE_MATRICULA, method = "pearson")
cor.test(data$VALOR_REPASSADO_EDUCACAO, data$MATRICULA, method = "pearson")
```
```{r}
# VALOR_REPASSADO_EDUCACAO

  ggqqplot(data$VALOR_REPASSADO_EDUCACAO, ylab = "Valor Repassado Educação") + 
    theme_classic()

```
```{r}
# IDEB_NOTA
  ggqqplot(data$IDEB_NOTA, ylab = "Nota Ideb Médi") + theme_classic()
```
```{r}
ggscatter(data, x = "IDEB_MEDIA", y = "VALOR_REPASSADO_EDUCACAO", 
          add = "reg.line", conf.int = TRUE, 
          cor.coef = TRUE, cor.method = "pearson",
          xlab = "Valor Repassado para Educação em Reais", ylab = "Nota ideb")
```

```{r}
ggscatter(data, x = "VALOR_REPASSADO_EDUCACAO", y = "IDH_EDUCACAO", 
          add = "reg.line", conf.int = TRUE, 
          cor.coef = TRUE, cor.method = "pearson",
          xlab = "Valor Repassado para Educação em Reais", ylab = "Nota ideb")
```
```{r}
ggscatter(data, x = "IDEB_MEDIA", y = "IDH_EDUCACAO", 
          add = "reg.line", conf.int = TRUE, 
          cor.coef = TRUE, cor.method = "pearson",
          xlab = "Valor Repassado para Educação em Reais", ylab = "Nota ideb")
```
Nesta visualização com `corrplot` não existe uma correlação forte do valor repassado para educação, com nenhuma outra variável, e exite uma boa correlação entre as notas do Ideb, e os medidores do IDH.

```{r}
correlations <- cor(data)
corrplot(correlations, method = "circle")
```
Já nesta visualização, podemos ver que existe uma correlação negativa baixa, dos valores repassados para educação, e a nota do Ideb.

```{r}
data_scale <- scale(data)
chart.Correlation(data_scale, histogram=TRUE)
```
Pelo boxplot, existem alguns potenciais outliers nas notas do Ideb.

```{r}
boxplot(data$IDEB_NOTA, main="Ideb Média")
```
```{r}
out_ideb <- boxplot.stats(data$IDEB_NOTA)$out
out_ideb_ind <- which(data$IDEB_NOTA %in% c(out_ideb))
data[out_ideb_ind, ]

```

```{r}
boxplot(data$IDH_EDUCACAO, main="IDH Educação")
```
```{r}
out_idh <-boxplot.stats(data$IDH_EDUCACAO)$out
out_idh_ind <- which(data$IDH_EDUCACAO %in% c(out_idh))
data[out_idh_ind, ]
```

```{r}
boxplot(data$REPASSE_MATRICULA, main="Valor Repassado Para Educação")
```

Essa visualização ficou meio estranha, não sei se teria que fazer algum ajuste:

```{r}
boxplot(data$VALOR_REPASSADO_EDUCACAO, main="Valor Repassado Para Educação")
```
```{r}
out_repasse <-boxplot.stats(data$VALOR_REPASSADO_EDUCACAO)$out
out_repasse_ind <- which(data$VALOR_REPASSADO_EDUCACAO %in% c(out_repasse))
data[out_repasse_ind, ]
```

```{r}
ggplot(data) +
  aes(x = VALOR_REPASSADO_EDUCACAO) +
  geom_histogram(bins = 30L, fill = "#0c4c8a") +
  theme_minimal()
```

```{r}
ggplot(data) +
  aes(x = IDEB_NOTA) +
  geom_histogram(bins = 30L, fill = "#0c4c8a") +
  theme_minimal()
```
```{r}
ggplot(data) +
    aes(x = IDH_EDUCACAO) +
  geom_histogram(bins = 30L, fill = "#0c4c8a") +
  theme_minimal()
```
